CherryPicked: [4.20] [Storage] remove artifactory usage in test_cdi_config.py
Refactors CDI config tests by moving them to tests/storage/cdi_config/, removing artifactory-dependent tests, and cleaning up unused fixtures.

Original PR: #4342
assisted by: claude code sonnet-4.5
